ShortPathing is now dynamic, and extra players added is calculated differently

+++ src/main/java/org/example/FlowEdge.java
@@ -33,7 +33,7 @@ public class FlowEdge extends DefaultWeightedEdge {
     //TODO Maybe implement dynamic calculation?
     // So far, this calculates the maximum delay possible
     public double calculateAggregateWeight() {
-        return ( this.queue.size() / this.capacity ) + this.actualWeight;
+        return ( ((double) this.queue.size()) / this.capacity ) + this.actualWeight;
         //return this.queue.size() + this.actualWeight;
     }
